requirements.txt
fastapi
uvicorn
pydantic
requests
PyMuPDF
python-docx
firebase-admin 
google-generativeai
PyPDF2
dotenv
stanza